Improvements: * ml_calcloss: added support for empty target/predictions * ml_trainproximal: cleaned up the way in which internal cv results are being summarized and how the best parameter setting is calculated * bci_train: can now pass in EvaluationScheme as 'off' (same as 0) * ParadigmCFS: no longer tries to locally pre-process data to get some model parameters (e.g., filter graph, chanlocs) * par_worker: now uses more retries when sending back, better logging * par_fileclient: now by default uses local fallback (also better logging) * utl_complete_model: now can accept a filter graph with a data set struct that has the data omitted (to reduce size) * utl_crossval/utl_collection_targets/utl_collection_tester: now supports case where no predictions get emitted * utl_collection_partition: removed unnecessary NoCrossvalidation mode (covered by bci_train) * liblsl-Matlab dependency: rebuilt some binaries * chopdeal: now supports arrays of more than 2bn entries
New features:
* parallel system can new schedule tasks larger than 2GB
(par_beginschedule, par_worker, Scheduler, ByteArrayMaker)
*

+++ dependencies/mexutils-2013-11-11/not_available_in_matlab/chopdeal.c
@@ -7,12 +7,12 @@
 void mexFunction(int nlhs, mxArray *plhs[],
                  int nrhs, const mxArray *prhs[])
 {
-    int datalen,numsizes,k,s,p;
+    long long datalen,numsizes,k,s,p;
     double *indata_d, *outdata_d;
     float *indata_f, *outdata_f;
     char *indata_c, *outdata_c;
     double *sizes;
-    int cursize;
+    long long cursize;
     mxArray *tmp;
     if (nrhs != 2)
         mexErrMsgIdAndTxt("chopdeal:incorrect_input_argument_number","2 input arguments required.");
